requirements.txt
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
yt-dlp>=2023.1.6
requests>=2.28.0
aiohttp>=3.8.0
mutagen>=1.46.0
ffmpeg-python>=0.2.0
pyyaml>=6.0
aiofiles>=0.8.0
pyexecjs>=1.5.1